+++ src/ConvDiffEq2D.h
@@ -11,12 +11,12 @@
 #include "DiffGrid2D.h"
 #include "Source.h"
 
-using Eigen::MatrixXd;
+using Eigen::MatrixXf;
 
 class ConvDiffEq2D : public Model2D
 {
     bool Record;
-    MatrixXd D, Vx, Vy, U_init;
+    MatrixXf D, Vx, Vy, U_init;
     PointSource2D source;
     DiffGrid2D U0, U1, DdxU, DdyU, VxU, VyU;
 
@@ -24,13 +24,13 @@ public:
     ConvDiffEq2D();
     ~ConvDiffEq2D() {}
 
-    void SetParameters(const MatrixXd& set_D, const MatrixXd set_Vx, const MatrixXd set_Vy);
-    void SetInit(const MatrixXd& set_U_init);
+    void SetParameters(const MatrixXf& set_D, const MatrixXf set_Vx, const MatrixXf set_Vy);
+    void SetInit(const MatrixXf& set_U_init);
     void SetSource(const PointSource2D& set_source);
     void SetRecord(bool r);
 
     void TimeUpdate(uint16_t idx_t);
     void Solve();
 
-    MatrixXd GetSol() const;
+    MatrixXf GetSol() const;
 };
